Firefox WebExtension Support - Part 2
