+++ ember-cli-build.js
@@ -164,12 +164,7 @@ module.exports = function(defaults) {
   });
 
   var minimumVersion = packageJson.emberVersionsSupported[0].replace(/\./g, '-');
-  var chromeRoot = 'panes-' + minimumVersion;
-  var firefoxRoot = 'data/' + chromeRoot;
-
-  var firefoxAndChromeExtra = new Funnel('shared', {
-    files: ['in-page-script.js'],
-  });
+  var webExtensionRoot = 'panes-' + minimumVersion;
 
   var replacementPattern = [{
     match: /{{env}}/,
@@ -187,13 +182,8 @@ module.exports = function(defaults) {
 
   replacementPattern = replacementPattern.concat(emberInspectorVersionPattern);
 
-  var skeletonChrome = replace('skeletons/chrome', {
+  var skeletonWebExtension = replace('skeletons/web-extension', {
     files: ['*'],
-    patterns: replacementPattern
-  });
-
-  var skeletonFirefox = replace('skeletons/firefox', {
-    files: ['*', '*/**'],
     patterns: replacementPattern
   });
 
@@ -203,17 +193,17 @@ module.exports = function(defaults) {
   });
 
   var firefox = mergeTrees([
-    mv(mergeTrees([tree, emberDebugs.firefox]), firefoxRoot),
-    skeletonFirefox
+    mv(mergeTrees([tree, emberDebugs.firefox]), webExtensionRoot),
+    skeletonWebExtension
   ]);
 
   var chrome = mergeTrees([
-    mv(mergeTrees([tree, emberDebugs.chrome]), chromeRoot),
-    skeletonChrome
+    mv(mergeTrees([tree, emberDebugs.chrome]), webExtensionRoot),
+    skeletonWebExtension
   ]);
 
   var bookmarklet = mergeTrees([
-    mv(mergeTrees([tree, emberDebugs.bookmarklet]), chromeRoot),
+    mv(mergeTrees([tree, emberDebugs.bookmarklet]), webExtensionRoot),
     skeletonBookmarklet
   ]);
 
@@ -227,7 +217,7 @@ module.exports = function(defaults) {
         bookmarklet
       ]);
       firefox = mergeTrees([
-        mv(prevDist + '/firefox/panes-' + version, 'data/panes-' + version),
+        mv(prevDist + '/firefox/panes-' + version, 'panes-' + version),
         firefox
       ]);
       chrome = mergeTrees([
@@ -238,20 +228,10 @@ module.exports = function(defaults) {
       var file = writeFile('index.html', "This Ember version is not supported in development environment.");
       var emberDebugFile = writeFile('ember_debug.js', 'void(0);');
       chrome = mergeTrees([mv(file, 'panes-' + version), chrome]);
-      firefox = mergeTrees([mv(file, 'data/panes-' + version), mv(emberDebugFile, 'data/panes-' + version), firefox]);
+      firefox = mergeTrees([mv(file, 'panes-' + version), firefox]);
       bookmarklet = mergeTrees([mv(file, 'panes-' + version), mv(emberDebugFile, 'panes-' + version), bookmarklet]);
     }
   });
-
-  firefox = mergeTrees([
-    mv(firefoxAndChromeExtra, 'data/scripts'),
-    firefox
-  ]);
-
-  chrome = mergeTrees([
-    mv(firefoxAndChromeExtra, 'scripts'),
-    chrome
-  ]);
 
   // Pass the current dist to the Ember Inspector app.
   // EMBER DIST
