+++ skeletons/web-extension/manifest.json
